backend/requirements.txt
fastapi==0.135.3
uvicorn==0.43.0
pdfplumber==0.11.9
anthropic==0.89.0
python-multipart==0.0.22
aiofiles==25.1.0
python-dotenv==1.2.2
